+++ vpin-studio-rest-client/src/main/java/module-info.java
@@ -13,4 +13,5 @@ open module de.mephisto.vpin.restclient {
   exports de.mephisto.vpin.restclient.client;
   exports de.mephisto.vpin.restclient.jobs;
   exports de.mephisto.vpin.restclient.popper;
+  exports de.mephisto.vpin.restclient.mame;
 }