+++ vpin-studio-server/src/main/java/de/mephisto/vpin/server/mame/MameResource.java
@@ -1,0 +1,25 @@
+package de.mephisto.vpin.server.mame;
+
+import de.mephisto.vpin.restclient.mame.MameOptions;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.*;
+
+import static de.mephisto.vpin.server.VPinStudioServer.API_SEGMENT;
+
+@RestController
+@RequestMapping(API_SEGMENT + "mame")
+public class MameResource {
+
+  @Autowired
+  private MameService mameService;
+
+  @GetMapping("/options/{rom}")
+  public MameOptions getOptions(@PathVariable("rom") String rom) {
+    return mameService.getOptions(rom);
+  }
+
+  @PostMapping("/options")
+  public MameOptions saveOptions(@RequestBody MameOptions options) {
+    return mameService.saveOptions(options);
+  }
+}