+++ RELEASE_NOTES.md
@@ -2,4 +2,4 @@
 
 - **Virtual Pinball Spreadsheet**: Added missing "open" link for the table selection.
 - **Virtual Pinball Spreadsheet**: Added update date to table selection.
-- 
+- **Preferences**: Added **VPin Mame** section.